Allow for providing uid/gid
The default uid/gid detection works fine if you're running on the same
host as the docker daemon, but that's not the case with boot2docker,
which provides another layer of indirection.

On my Mac, my uid/gid is 600/20.  In the boot2docker vm, the docker
user is 1000/50.  When using VirtualBox to share files from my Mac's
home directory to the VM (and hence to Docker) the value from
os.getuid() doesn't match the owning uid of some files, so rpmbuild
fails.